requirements.txt

Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.1
opencv-python==4.5.2.54
Pillow==8.2.0
psycopg2-binary==2.8.6



